refactor server code (Task #72)
